Chris

+++ src/main/java/com/gl05/bad/dao/EscuelaPostgradoDao.java
@@ -3,6 +3,6 @@ package com.gl05.bad.dao;
 import com.gl05.bad.domain.EscuelaPostgrado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EscuelaPostgradoDao extends JpaRepository<EscuelaPostgrado, Integer>{
+public interface EscuelaPostgradoDao extends JpaRepository<EscuelaPostgrado, Long>{
     
 }
